cron/requirements.txt
pymongo==4.6.1
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
python-dotenv==1.0.0
